+++ lib/zaikio/hub/vendor.rb
@@ -1,0 +1,10 @@
+module Zaikio
+  module Hub
+    class Vendor < Base
+      uri nil
+
+      # Attributes
+      attributes :name, :slug, :logo_url
+    end
+  end
+end